moves map out of its own folder (due to it being alone)
