+++ src/main/java/app/controller/graphicsEngine/GraphicsEngine.java
@@ -1,7 +1,7 @@
 package app.controller.graphicsEngine;
 
 import app.model.agents.Agent;
-import app.model.map.Map;
+import app.model.Map;
 
 import java.util.ArrayList;
